Reduces the maximum cli/srv score per callback to 128

+++ include/Flow.h
@@ -280,7 +280,7 @@ class Flow : public GenericHashEntry {
     - Synchronous:  The alerts bitmap is updated and the predominant alert is possibly updated.
                     Immediate alert JSON generation and enqueue to the recipients are performed as well.
    */
-  bool setAlertsBitmap(FlowAlertType alert_type, AlertLevel alert_severity, u_int16_t cli_inc, u_int16_t srv_inc, bool async);
+  bool setAlertsBitmap(FlowAlertType alert_type, AlertLevel alert_severity, u_int8_t cli_inc, u_int8_t srv_inc, bool async);
   void setNormalToAlertedCounters();
 
  public:
@@ -311,13 +311,13 @@ class Flow : public GenericHashEntry {
     cause the alert JSON to be generated after the call.
     The FlowCallback should implement the buildAlert() method which is called in the predominant callback to actually build the FlowAlert object.
    */
-  bool triggerAlertAsync(FlowAlertType alert_type, AlertLevel alert_severity, u_int16_t cli_score_inc, u_int16_t srv_score_inc);
+  bool triggerAlertAsync(FlowAlertType alert_type, AlertLevel alert_severity, u_int8_t cli_score_inc, u_int8_t srv_score_inc);
 
   /* 
      Called by FlowCallback subclasses to trigger a flow alert. This is a syncrhonous call, more expensive, but
      causes the alert (FlowAlert) to be immediately enqueued to all recipients.
    */
-  bool triggerAlertSync(FlowAlert *alert, AlertLevel alert_severity, u_int16_t cli_score_inc, u_int16_t srv_score_inc);
+  bool triggerAlertSync(FlowAlert *alert, AlertLevel alert_severity, u_int8_t cli_score_inc, u_int8_t srv_score_inc);
   /*
     Enqueues the predominant alert of the flow to all available flow recipients.
    */
