+++ include/flow_callbacks/FlowRiskHTTPNumericIPHost.h
@@ -30,8 +30,8 @@ class FlowRiskHTTPNumericIPHost : public FlowRisk {
   FlowAlertType getAlertType() const { return FlowRiskHTTPNumericIPHostAlert::getClassType(); };
 
   /* Uncomment to override the default scores:
-  u_int16_t getClientScore() { return 50; }
-  u_int16_t getServerScore() { return 50; }
+  u_int8_t getClientScore() { return 50; }
+  u_int8_t getServerScore() { return 50; }
   */
 
  public:
