+++ include/flow_callbacks/FlowRiskSMBInsecureVersion.h
@@ -30,8 +30,8 @@ class FlowRiskSMBInsecureVersion : public FlowRisk {
   FlowAlertType getAlertType() const { return FlowRiskSMBInsecureVersionAlert::getClassType(); }
 
   /* Uncomment to override the default scores:
-  u_int16_t getClientScore() { return 50; }
-  u_int16_t getServerScore() { return 50; }
+  u_int8_t getClientScore() { return 50; }
+  u_int8_t getServerScore() { return 50; }
   */
 
  public:
