+++ include/flow_callbacks/FlowRiskURLPossibleSQLInjection.h
@@ -30,8 +30,8 @@ class FlowRiskURLPossibleSQLInjection : public FlowRisk {
   FlowAlertType getAlertType() const { return FlowRiskURLPossibleSQLInjectionAlert::getClassType(); }
 
   /* Uncomment to override the default scores:
-  u_int16_t getClientScore() { return 50; }
-  u_int16_t getServerScore() { return 50; }
+  u_int8_t getClientScore() { return 50; }
+  u_int8_t getServerScore() { return 50; }
   */
 
  public:
