+++ src/flow_callbacks/ExternalAlertCheck.cpp
@@ -24,8 +24,8 @@
 
 void ExternalAlertCheck::checkExternalAlert(Flow *f) {
   if (f->hasExternalAlert()) { 
-    u_int16_t c_score = 100;
-    u_int16_t s_score = 100;
+    u_int8_t c_score = 100;
+    u_int8_t s_score = 100;
 
     f->triggerAlertAsync(ExternalAlertCheckAlert::getClassType(), getSeverity(), c_score, s_score);
   }
