+++ src/flow_callbacks/NotPurged.cpp
@@ -26,7 +26,7 @@
 
 void NotPurged::checkNotPurged(Flow *f) {
   if(f->isNotPurged()) {
-    u_int16_t c_score = 0, s_score = 0;
+    u_int8_t c_score = 0, s_score = 0;
 
     f->triggerAlertAsync(NotPurgedAlert::getClassType(), getSeverity(), c_score, s_score);
   }
