+++ src/flow_callbacks/TCPZeroWindow.cpp
@@ -24,7 +24,7 @@
 
 void TCPZeroWindow::checkTCPWindow(Flow *f) {
   if(f->isTCPZeroWindow()) {
-    u_int16_t c_score = 30, s_score = 30;
+    u_int8_t c_score = 30, s_score = 30;
 
     f->triggerAlertAsync(TCPZeroWindowAlert::getClassType(), getSeverity(), c_score, s_score);
   }
