+++ src/flow_callbacks/UnexpectedServer.cpp
@@ -97,7 +97,7 @@ void UnexpectedServer::protocolDetected(Flow *f) {
   if(!isAllowedProto(f)) return;
   
   if(!isAllowedHost(getServerIP(f))) {
-    u_int16_t c_score = 100, s_score = 100;
+    u_int8_t c_score = 100, s_score = 100;
     
     f->triggerAlertAsync(getAlertType(), getSeverity(), c_score, s_score);   
   }
