+++ src/flow_callbacks/WebMining.cpp
@@ -24,7 +24,7 @@
 
 void WebMining::protocolDetected(Flow *f) {
   if (f->get_protocol_category() == CUSTOM_CATEGORY_MINING) {
-    u_int16_t c_score = 50, s_score = 10;
+    u_int8_t c_score = 50, s_score = 10;
    
     f->triggerAlertAsync(WebMiningAlert::getClassType(), getSeverity(), c_score, s_score);
   }
